+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 2.0.0-alpha.3
 * Added `<Id, Data>` type parameters to `Service`.
 * `HookedService` now follows suit, and takes a third parameter, pointing to the inner service.
+* `Routable.use` now uses the generic parameters added to `Service`.
+* Added generic usage to `HookedServiceListener`, etc.
 
 # 2.0.0-alpha.2
 * Added `ResponseContext.detach`.